Tenative plot of G1 trajectory
getting familiar with Galileo trajectory data by plotting it against Ganymede. there are 2 versions of this code -  the graph w/units of Ganymede radii and one with units of Jupiter radii. the data says it was in terms of Jupiter radii, but it only made sense dimensionally in terms of Ganymede radii. will double check which is correct.